refactor: rename `Eval` -> `EvalFull`
