+++ src/primer-api/index.ts
@@ -1,3 +1,2 @@
 export * from './primer-api';
 export * from './model';
-
